+++ firmware/src/serial.c
@@ -130,11 +130,16 @@ void serial_write(uint8_t data) {
 inline void serial_write_number(double num) {
   // num to be [-134217.728, 134217.727]
   // three decimals are retained
-  uint32_t numint = lround(num*1000 + 134217728);
-  serial_write((numint&127)+128);
-  serial_write(((numint&((uint32_t)127<<7))>>7)+128);
-  serial_write(((numint&((uint32_t)127<<14))>>14)+128);
-  serial_write(((numint&((uint32_t)127<<21))>>21)+128);
+  int32_t numint = lround(num*1000.0) + 134217728L;
+  serial_write((numint&127UL)+128);
+  serial_write(((numint&(127UL<<7))>>7)+128);
+  serial_write(((numint&(127UL<<14))>>14)+128);
+  serial_write(((numint&(127UL<<21))>>21)+128);
+  // uint32_t numint = lround(num*1000 + 134217728UL);
+  // serial_write((numint&127)+128);
+  // serial_write(((numint&(127UL<<7))>>7)+128);
+  // serial_write(((numint&(127UL<<14))>>14)+128);
+  // serial_write(((numint&(127UL<<21))>>21)+128);
 }
